requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
yt-dlp==2024.8.6
ffmpeg-python==0.2.0
Pillow==10.4.0
